edit pointer function

+++ main.h
@@ -22,7 +22,7 @@ int print_binary(va_list arg);
 int print_hex(va_list arg, int flag);
 int print_HEX(va_list arg, int flag);
 int print_HEX2(unsigned int num);
-int print_hex_to_pointer(unsigned int num);
+int print_hex_to_pointer(unsigned long int num);
 int print_octal(va_list arg);
 
 /*Helper Fnction*/
